pm25/1st Place/requirements.txt
loguru==0.6.0
scikit-learn==1.0.2
pandas
numpy
matplotlib
pyaml==21.10.1
opencv-python==4.5.5.62
xgboost==1.5.2
catboost==1.0.4
lightgbm==3.3.2
seaborn==0.11.2
pygrib==2.1.4
tqdm==4.62.3
awscli
scipy
netCDF4==1.5.8
rioxarray
requests
jupyter
jupyterlab